Create Russian Kaggle Project
This is the initial data load, it also updates all Null values with median values from the data set. Then it goes on to start cleansing the Built_Year column. I will continue to work on this on my end, I think it will be important to break down the data similar to the way that is in the videos were we are looking at continuous variables together and so on.